+++ packages/core/CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Update `@tvkitchen/base-classes` to version `2.0.0-alpha.1`.
 - Update `@tvkitchen/base-interfaces` to version `4.0.0-alpha.4`.
+- Update `AbstractVideoIngestionAppliance` to populate Payload `origin` instead of `timestamp`.
+- Remove arrow function method definitions from `AbstractVideoIngestionAppliance` and `AbstractAppliance`.
 
 ## [0.6.1] - 2020-04-07
 ### Added