+++ packages/core/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.7.0] - 2020-05-13
 ### Added
 - `AbstractAppliance` will now populate the `origin` of outbound payloads.
 - Update `AbstractVideoIngestionAppliance` to filter corrupt payloads.
@@ -73,8 +75,9 @@ This package is identical to [0.6.1] but was deprecated due to a publication err
 - Increase the `base-classes` version dependency to `1.3.0`.
 - Increase the `base-interfaces` version dependency to `3.0.0`.
 
-[Unreleased]: https://github.com/tvkitchen/appliances/compare/@tvkitchen/appliance-video-file-ingestion@0.6.0...HEAD
-[0.6.0]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-core@0.6.0
+[Unreleased]: https://github.com/tvkitchen/appliances/compare/@tvkitchen/appliance-video-file-ingestion@0.7.0...HEAD
+[0.7.0]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-core@0.7.0
+[0.6.1]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-core@0.6.1
 [0.5.0]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-core@0.5.0
 [0.4.0]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-core@0.4.0
 [0.3.1]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-core@0.3.1